fix decoration bug ; organize
